requirements.txt
# Core test framework
flask>=2.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-html>=4.0.0

# Load testing
locust>=2.17.0

# Utilities
jsonschema>=4.19.0
requests>=2.31.0
bandit>=1.7.5
python-dotenv>=1.0.0

# Real-service integration tests (requires docker-compose up -d)
pika>=1.3.0
psycopg2-binary>=2.9.0
